extra

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	long int i, j, k, l;
+	long long int i, j, k, l;
 
 	j = 1;
 	k = 2;
